Add the rules used for importing amarok and phonon
Thiago wrote;
Please note that the scripts I sent last year do not take into account the
modifications I made to svn2git to do annotated tagging. The Amarok and Phonon
rules do, so please update all my old rules to create tags properly.

So, here are the fixed rules (obsolete for the actual import) which would
be a great way to show the correct way of working.