Add the commit info for all broke tags.
Most of the broken ones are due to cvs2svn, I've taken the highest revision
number which was added to the tag and belonged to the kdevelop sources.
After the updates to the matching rules all of these commits should be
found in the git log of the generated git repositories, if not there's an
error somewhere.