Fix rules for kdevelop-pg-qt.
There was a commit to the kdevelop-pg directory after the fork which
screwed the history and also the code itself. So better stop taking
commits from the kdevelop-pg directory after the fork.

Also stop at the last svn rev which had content for kdevelop-pg-qt so
we don't include the removal and replacing with a README commit.